Add files via upload
This script is a collection of various other script, but in this case each script has its own separate folder and files.

The input is a domain, then the we check this domain on kwbreaker script to get its term. Then we check the term on various services. Output is a combination of the scraped data.

The services this scrip is scraping are:

Google Search using Serper API.
NP - https://projects.propublica.org/nonprofits/
TM (to be replaced by WIPO)
OC - OpenCorporates.com